shared/imgproc: fixed grayd_pixel_t typedef (it's double, not float)
git-svn-id: https://scm.citationtech.net/trunk/shared/imgproc/imgproc@10509 b7e6147c-0f76-4c84-a09d-19069afaf10c

+++ imgproc/gil-float-image.hpp
@@ -17,7 +17,7 @@ typedef image<grayf_pixel_t, false> grayf_image_t;
 typedef grayf_image_t::view_t grayf_view_t;
 typedef grayf_image_t::const_view_t grayf_const_view_t;
 
-typedef pixel<float, gray_layout_t> grayd_pixel_t;
+typedef pixel<double, gray_layout_t> grayd_pixel_t;
 typedef image<grayd_pixel_t, false> grayd_image_t;
 typedef grayd_image_t::view_t grayd_view_t;
 typedef grayd_image_t::const_view_t grayd_const_view_t;
